- replaced older YAPE point (yape::perform_one_point()) with newer, nicer one from OpenCV::LDetector (yape::perform_one_point_2(), newer version of YAPE). LDetect code is provided for reference in cvplanardetect.cpp/.h but not compiled into the project - multithreaded yape::raw_detect (see yape::raw_detect_mt()) - scale fix to FProfiler timing code - exposed tau setting for yape, changed default from 10 to 25
--> please re-train existing classifiers

	modified:   artvertiser.cbp
	modified:   artvertiser.layout
	modified:   artvertiser/FProfiler/FProfiler.cpp
	modified:   artvertiser/FProfiler/FProfiler.h
	modified:   artvertiser/FProfiler/FSemaphore.h
	modified:   artvertiser/FProfiler/FTime.h
	modified:   artvertiser/artvertiser.cpp
	modified:   garfeild/keypoints/yape.cpp
	modified:   garfeild/keypoints/yape.h
	new file:   garfeild/viewsets/cvplanardetect.cpp
	new file:   garfeild/viewsets/cvplanardetect.h
	modified:   garfeild/viewsets/planar_object_recognizer.cpp
	modified:   garfeild/viewsets/planar_object_recognizer.h

+++ artvertiser/FProfiler/FTime.h
@@ -27,7 +27,7 @@
 class FTime
 {
 public:
-	FTime() { 
+	FTime() {
 		#ifdef OSX
 		time = 0;
 		#else
@@ -41,10 +41,10 @@ public:
 
 	// update our time to the current time, returning
 	// delta time in seconds as a float
-	float Update();
+	double Update();
 
 	// get the last delta time returned by Update()
-	float GetLastUpdateTime() { return last_update_time; }
+	double GetLastUpdateTime() { return last_update_time; }
 
 	// return seconds
 	double ToMillis();
@@ -59,7 +59,7 @@ private:
 	#else
 	timespec time;
 	#endif
-	float last_update_time;
+	double last_update_time;
 
 };
